Move `APIConfig` class to `cli._config`
Keep the config stuff together
